# docs/openapi_current.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "bin2nlp API",
        "description": "\n        ## Binary Decompilation & Multi-LLM Translation Service\n\n        Transform binary executables into human-readable analysis through advanced decompilation \n        and multi-provider LLM translation.\n\n        ### Key Features\n        - **Multi-format support**: PE, ELF, Mach-O binary analysis\n        - **Multi-LLM providers**: OpenAI, Anthropic, Google Gemini integration\n        - **Scalable architecture**: PostgreSQL database, file storage, async processing\n        - **Production ready**: Open access, monitoring, observability\n\n        ### Open Access\n        All endpoints are freely accessible without authentication requirements.\n        This service is designed for open binary analysis and research.\n\n        ### Support\n        - Documentation: [GitHub Repository](https://github.com/example/bin2nlp)\n        - Issues: [Issue Tracker](https://github.com/example/bin2nlp/issues)\n        ",
        "termsOfService": "https://example.com/terms",
        "contact": {
            "name": "bin2nlp Support",
            "url": "https://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.bin2nlp.com",
...
        }
    ],
    "paths": {
...
        "/api/v1/llm-providers/{provider_id}": {
            "get": {
                "tags": [
                    "llm-providers"
                ],
                "summary": "Get Llm Provider Details",
                "description": "Get detailed information about a specific LLM provider.\n\nReturns comprehensive information including health status,\nrecent performance metrics, and configuration details.",
                "operationId": "get_llm_provider_details_api_v1_llm_providers__provider_id__get",
                "parameters": [
                    {
                        "name": "provider_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Provider Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}